Restructure concrete_async
